+++ airline-gateway/src/main/java/app/airlinegateway/security/config/EndpointConfig.java
@@ -2,8 +2,10 @@ package app.airlinegateway.security.config;
 
 import lombok.Data;
 
+import java.util.List;
+
 @Data
 public class EndpointConfig {
     String uri;
-    String method;
+    List<String> methods;
 }